Update readme.md
The full data needed for my project will not be ready/available before the end of April 2023. To overcome this challenge, I will add all the currently available data as much as possible and fill the gap with some dummy data that helps me build the dashboard until the real data is available. I hope this is fine by you.